pyproject.toml
[tool.poetry]
name = "forumbot"
version = "0.1.0"
description = ""
authors = ["Sören Wegener <soerface@flipdot.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydiscourse = "1.1.1"
requests = "^2.32.3"
schedule = "0.6.0"
Jinja2 = "^3.1.4"
PyYAML = "^6.0.2"
numpy = "^2.1.2"
mistune = "^3.0.2"
MarkupSafe = "^3.0.2"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
sentry-sdk = "^2.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
